consumer/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
